Start of merge from 2.2. Added locking fixes. Now for the delete on close bug. Jeremy.

+++ source/locking/posix.c
@@ -25,7 +25,7 @@
 
 #include "includes.h"
 extern int DEBUGLEVEL;
-extern int global_smbpid;
+extern uint16 global_smbpid;
 
 /*
  * The POSIX locking database handle.
